Create grading.py
In this example, the get_score function is used to prompt the user to enter their exam score. The get_grade function takes the score as input and uses an if-elif statement to determine the corresponding grade, which is then returned. The grade is then printed to the screen along with the message "Your grade is: