Create magic-8-ball.py
n this example, a list of possible responses for the Magic 8 Ball is created and stored in the responses variable. The magic_8_ball function uses the random.choice function from the random module to randomly choose one of the responses and return it. Finally, the returned response is printed to the screen.