Create random-password-generator.py
In this example, the generate_password function takes an optional parameter length which specifies the length of the password to be generated (default is 8 characters). The function uses the random.choices function from the random module to randomly choose length number of characters from a combination of letters, digits, and punctuation. The returned password is a string created by joining these randomly chosen characters. Finally, the returned password is printed to the screen.