Create dictionaries.py
In this example, a dictionary named person is created, where the keys are "name", "age", and "email" and their respective values. The values in a dictionary can be accessed using square bracket notation, such as person["name"], which would return "John Doe". The values in a dictionary can also be updated by assigning a new value to a key, such as person["age"] = 33. New key-value pairs can be added to a dictionary using square bracket notation, such as person["address"] = "123 Main St". Finally, key-value pairs can be removed from a dictionary using the del statement, such as del person["email"]. Note that the person.get() method can be used to access the value of a key in a dictionary, but if the key does not exist it will return a default value (in this case "Not found").