Create average-numbers.py
In this code, the list of numbers is stored in the numbers variable. A for loop is used to iterate through the list and add up the elements, storing the result in the sum variable. The length of the list is then obtained using the len function, and the average is calculated by dividing sum by len(numbers). Finally, the average is printed to the screen.