Create guess-the-number.py
In this code, the random module is imported and used to generate a random integer between 1 and 10. A while loop is used to repeatedly prompt the user for a guess and check if it is correct. The user's guess is obtained using the input function and converted to an integer. If the guess is correct, the program prints a message and breaks out of the loop. If the guess is too low, the program prints a message indicating that the guess is too low. If the guess is too high, the program prints a message indicating that the guess is too high.