Create image-manipulation.py
This code first opens the image file using the Image.open() function from the Pillow library. Then, it uses the convert() method to convert the image to greyscale. The parameter "L" specifies that the image should be converted to an 8-bit greyscale image. The code then saves the converted image using the save() method and shows the image using the show() method.