Create rain-data.py
This code makes a request to the OpenWeatherMap API to retrieve the current weather information for Las Vegas, NV, USA. The temperature is returned in Kelvin, so we convert it to Celsius. Note that you will need to sign up for a free API key from OpenWeatherMap in order to use this code.